fixed editor

+++ app/Http/Controllers/PostController.php
@@ -3,9 +3,21 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
+
 class PostController extends Controller
 {
+     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -39,13 +51,33 @@ class PostController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
-        
+        // Handle file upload
+       if($request->hasFile('cover_image')) {
+        //get file name
+        $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+        //get just filename
+        $filename = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+        //get extension
+        $extension = $request->file('cover_image')->getClientOriginalExtension();
+    //filename to store
+        $fileNameToStore = $filename . '_' . time() . '.' . $extension;
+       //upload image
+        $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+    } else {
+           $fileNameToStore = 'noimage.png';
+    }
+
+        //create post
+
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id; 
+        $post->cover_image = $fileNameToStore; 
+
         $post->save();
         return redirect('/posts')->with('success', 'Post Created');
     }
@@ -71,8 +103,11 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        // Check for correct user 
+        if(auth()->user()->id != $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
         return view('posts.edit')->with('post', $post);
-        //return $id;
     }
 
     /**
@@ -89,9 +124,26 @@ class PostController extends Controller
             'body' => 'required'
         ]);
         
+         // Handle file upload
+       if($request->hasFile('cover_image')) {
+            //get file name
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //get just filename
+            $filename = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            //get extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+        //filename to store
+            $fileNameToStore = $filename . '_' . time() . '.' . $extension;
+        //upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } 
+
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image')) {
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
         return redirect('/posts')->with('success', 'Post Updated');
     }
@@ -105,6 +157,14 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+         // Check for correct user 
+         if(auth()->user()->id != $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
+
+        if($post->cover_image != 'noimage.png') {
+            Storage::delete('public/cover_images/' . $post->cover_image);
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post Removed');
     }
